Move CLI workflow helpers under workflow
